Refactor canonicalization internals and remove benchmark scaffolding
