+++ Cargo.toml
@@ -10,7 +10,6 @@ readme = "README.md"
 
 [features]
 default = []
-benchmarking = []
 fuzzing = []
 
 [dependencies]
@@ -44,7 +43,6 @@ invalid_rust_codeblocks = "forbid"
 unreadable_literal = "allow"
 missing_panics_doc = "allow"
 new_ret_no_self = "allow"
-struct_excessive_bools = "allow"
 should_panic_without_expect = "allow"
 doc_markdown = "allow"
 clone_on_copy = "forbid"
@@ -55,19 +53,9 @@ pedantic = { level = "deny", priority = -1 }
 lto = "thin"
 codegen-units = 1
 
-[profile.bench]
-inherits = "release"
-
 [dev-dependencies]
-criterion = "0.8.2"
-csv = "1.4.0"
 flate2 = "1.1.9"
 indicatif = "0.18.4"
 rayon = "1.11.0"
-reqwest = { version = "0.12.24", default-features = false, features = ["blocking", "rustls-tls"] }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
-
-[[bench]]
-name = "smiles_bench"
-harness = false
