+++ src/smiles/canonicalization/support/mod.rs
@@ -1,0 +1,6 @@
+mod common;
+mod invariants;
+
+#[cfg(test)]
+pub(super) use common::{permute_smiles, same_canonicalization_state};
+pub(super) use invariants::assert_canonicalization_invariants;
