+++ src/smiles/canonicalization/tests/mod.rs
@@ -1,0 +1,3 @@
+mod exact_and_rewrite;
+mod regressions;
+mod stereo_distinctions;
