+++ src/smiles/canonicalization/tests/regressions.rs
@@ -1,0 +1,144 @@
+use core::str::FromStr;
+
+use geometric_traits::traits::SparseValuedMatrixRef;
+
+use super::super::{
+    Smiles, canonicalization_state_key,
+    support::{assert_canonicalization_invariants, same_canonicalization_state},
+};
+
+fn hidden_bond_order_digest(smiles: &Smiles) -> u64 {
+    use core::hash::{Hash, Hasher};
+    use std::collections::hash_map::DefaultHasher;
+
+    let mut hasher = DefaultHasher::new();
+    smiles
+        .bond_matrix()
+        .sparse_entries()
+        .filter(|((row, column), _)| row < column)
+        .for_each(|((row, column), entry)| (row, column, entry.order()).hash(&mut hasher));
+    hasher.finish()
+}
+
+#[test]
+fn canonicalize_handles_wildcard_quadruple_bond_regression() {
+    let smiles = Smiles::from_str("*$*").unwrap();
+
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_handles_aromatic_sulfur_chain_regression() {
+    let smiles = Smiles::from_str("scCC").unwrap();
+
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_handles_mixed_aromatic_phosphorus_wildcard_regression() {
+    let smiles = Smiles::from_str("bsC3P1*sC3P1**p1C3P1**OC3p1C3Os*p1C3P1**OC3p1C3Os").unwrap();
+
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_handles_polywildcard_aromatic_oxygen_regression() {
+    let smiles =
+        Smiles::from_str("OO*c8Oc8OOOO**c8OOO*c8OOOO*CO*c8OOOO*c8OOOO**c8OOO*c8OOOO*COOOO")
+            .unwrap();
+
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_handles_aromatic_wildcard_triangle_regression() {
+    let smiles = Smiles::from_str("P1*P1").unwrap();
+
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_handles_fuzz_crash_dfd86928_regression() {
+    let smiles =
+        Smiles::from_str("CBBBBBBBBBBBBBBBBBBBBBBB.OBBBBBBBBBBBBBBBBBBBB.OBBP1*P1**BBBBBBB")
+            .unwrap();
+
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_handles_fuzz_crash_partial_aromatic_diagnostics_regression() {
+    let smiles =
+        Smiles::from_str("bsp1C2P1**C3o1**p1*CPp*O14C3Po1**p1C2P1**C3o1$*p1C4P1**Op1C3Os").unwrap();
+
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_handles_fuzz_crash_component_order_budget_regression() {
+    let smiles = Smiles::from_str("Oooooooooooooooooooo.****.*ccOc**S*nnbsC3P1*soooooooooo.****.*ccOc**S*nnbsC3P1*sC3P1**p1C3P1**OC3p1C3Os*p1C3P2***S*nnnCnsnC3P1**p1C3P1**OC3p1C3Os*p1C3P2***S*nnnCns").unwrap();
+
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_handles_fuzz_crash_new_regression_candidate() {
+    let smiles = Smiles::from_str("***SCbs=3P1o*sC3P1**p1C3P1**OC3P1*=*OC3p1C3Os").unwrap();
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_handles_fuzz_crash_db6568a7_regression() {
+    let smiles = Smiles::from_str(
+        "*[*@]*.**.******[*@]*.**.**.*.**.**.******.**.****.*.*.**.**.***/**.**.****.**.***.**.**.*.**.**.******N**.****.**.**.**",
+    )
+    .unwrap();
+
+    assert_canonicalization_invariants(&smiles);
+}
+
+#[test]
+fn canonicalize_reaches_fixed_point_for_pubchem_cid_10888631_regression() {
+    let smiles =
+        Smiles::from_str("C1C2[C@@H]3[C@H]4[C@H]5[C@H]6C5CC4[C@@H]6[C@H]2[C@@H]7[C@@H]3C71")
+            .unwrap();
+
+    let canonicalized = smiles.canonicalize();
+    assert!(canonicalized.is_canonical(), "canonicalize() did not reach a canonical fixed point");
+
+    let recanonicalized = canonicalized.canonicalize();
+    same_canonicalization_state(&canonicalized, &recanonicalized);
+}
+
+#[test]
+fn canonicalize_reaches_fixed_point_for_pubchem_cid_101510359_regression() {
+    let smiles = Smiles::from_str("CC(C)(C)C1=CC(=CC(=C1)C2=C3C=CC(=C(C4=NC5=CC6=C(C=C(N6)C(=C7C=CC2=N7)C8=CC(=CC(=C8)C(C)(C)C)C(C)(C)C)C9=C1C=CC(=C(C2=NC(=C(C6=CC=C(N6)C(=C6C=CC9=N6)C6=CC(=CC(=C6)C(C)(C)C)C(C)(C)C)C6=CC(=CC(=C6)C(C)(C)C)C(C)(C)C)C=C2)C2=C6C=C7C(=CC(=N7)C(=C7C=CC(=C(C8=NC(=C(C(=C2)N6)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)C=C8)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)N7)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)C2=C6C=CC(=C(C5=C4)C4=NC(=C(C5=CC=C(N5)C(=C5C=CC2=N5)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)C=C4)N6)N1)C1=CC(=CC(=C1)C(C)(C)C)C(C)(C)C)N3)C(C)(C)C").unwrap();
+
+    let canonicalized = smiles.canonicalize();
+    let recanonicalized = canonicalized.canonicalize();
+
+    same_canonicalization_state(&canonicalized, &recanonicalized);
+}
+
+#[test]
+fn canonicalization_spelling_normal_form_clears_implicit_hydrogen_cache() {
+    let aromaticized = Smiles::from_str("C1=CC=CC=C1")
+        .unwrap()
+        .perceive_aromaticity()
+        .unwrap()
+        .into_aromaticized();
+
+    assert!(aromaticized.implicit_hydrogen_cache.is_some());
+    assert!(aromaticized.canonicalization_spelling_normal_form().implicit_hydrogen_cache.is_none());
+}
+
+#[test]
+fn canonicalization_step_does_not_hide_bond_order_drift_for_pubchem_cid_101510359() {
+    let smiles = Smiles::from_str("CC(C)(C)C1=CC(=CC(=C1)C2=C3C=CC(=C(C4=NC5=CC6=C(C=C(N6)C(=C7C=CC2=N7)C8=CC(=CC(=C8)C(C)(C)C)C(C)(C)C)C9=C1C=CC(=C(C2=NC(=C(C6=CC=C(N6)C(=C6C=CC9=N6)C6=CC(=CC(=C6)C(C)(C)C)C(C)(C)C)C6=CC(=CC(=C6)C(C)(C)C)C(C)(C)C)C=C2)C2=C6C=C7C(=CC(=N7)C(=C7C=CC(=C(C8=NC(=C(C(=C2)N6)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)C=C8)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)N7)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)C2=C6C=CC(=C(C5=C4)C4=NC(=C(C5=CC=C(N5)C(=C5C=CC2=N5)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)C2=CC(=CC(=C2)C(C)(C)C)C(C)(C)C)C=C4)N6)N1)C1=CC(=CC(=C1)C(C)(C)C)C(C)(C)C)N3)C(C)(C)C").unwrap();
+
+    let entry = smiles.canonicalize();
+    let direct_step = entry.canonicalization_step();
+
+    assert_eq!(canonicalization_state_key(&entry), canonicalization_state_key(&direct_step));
+    assert_eq!(hidden_bond_order_digest(&entry), hidden_bond_order_digest(&direct_step));
+}
